+++ apps/www/src/app/(main)/(docs)/components/react-hook-form/examples/editor.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import { zodResolver } from "@hookform/resolvers/zod";
-import { RHFEditor } from "@jamsr-ui/react";
 import { useForm } from "react-hook-form";
 import { any, object, record } from "zod";
 import { RHFDemoWrapper } from "../components/wrapper";
+import { RHFEditor } from "@jamsr-ui/rhf";
 
 type FormValues = {
   description: Record<string, unknown>;