+++ apps/www/src/app/(main)/(docs)/components/react-hook-form/examples/file-upload-multi.tsx
@@ -4,11 +4,8 @@ import { type ImageMetadata } from "@/app/config";
 import { CDN_API_URL, CDN_UPLOAD_URL } from "@/utils/config";
 import { zodImage } from "@/utils/zod";
 import { zodResolver } from "@hookform/resolvers/zod";
-import {
-  type FileUploadError,
-  RHFFileUploadMulti,
-  toast,
-} from "@jamsr-ui/react";
+import { type FileUploadError, toast } from "@jamsr-ui/react";
+import { RHFFileUploadMulti } from "@jamsr-ui/rhf";
 import { useForm } from "react-hook-form";
 import { array, object } from "zod";
 import { RHFDemoWrapper } from "../components/wrapper";