+++ apps/www/src/app/(main)/(docs)/components/react-hook-form/examples/file-upload-single-default-value.tsx
@@ -4,14 +4,11 @@ import { type ImageMetadata } from "@/app/config";
 import { CDN_API_URL, CDN_UPLOAD_URL } from "@/utils/config";
 import { zodImage } from "@/utils/zod";
 import { zodResolver } from "@hookform/resolvers/zod";
-import {
-  type FileUploadError,
-  RHFFileUploadSingle,
-  toast,
-} from "@jamsr-ui/react";
+import { type FileUploadError, toast } from "@jamsr-ui/react";
 import { useForm } from "react-hook-form";
 import { object } from "zod";
 import { RHFDemoWrapper } from "../components/wrapper";
+import { RHFFileUploadSingle } from "@jamsr-ui/rhf";
 
 type FormValues = {
   image: ImageMetadata;