+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ui/storybook
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/react@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes