+++ packages/components/rhf/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/rhf
 
+## 0.1.1
+
+### Patch Changes
+
+- refactor: rhf
+
 ## 0.1.0
 
 ### Minor Changes